// src/frontend/src/backend.d.ts
import type { Principal } from "@icp-sdk/core/principal";
export interface Some<T> {
    __kind__: "Some";
    value: T;
}
export interface None {
    __kind__: "None";
}
export type Option<T> = Some<T> | None;
export class ExternalBlob {
    getBytes(): Promise<Uint8Array<ArrayBuffer>>;
    getDirectURL(): string;
    static fromURL(url: string): ExternalBlob;
    static fromBytes(blob: Uint8Array<ArrayBuffer>): ExternalBlob;
    withUploadProgress(onProgress: (percentage: number) => void): ExternalBlob;
}
export interface ContactSubmission {
    id: string;
    name: string;
    email: string;
    message: string;
    timestamp: bigint;
    phone?: string;
}
export interface Service {
    id: string;
    title: string;
    order: bigint;
    description: string;
    photoUrl?: ExternalBlob;
    iconName: string;
    shortDescription: string;
    videoUrl?: ExternalBlob;
}
export interface UserProfile {
    name: string;
}
export enum UserRole {
    admin = "admin",
    user = "user",
    guest = "guest"
}
export interface backendInterface {
    addService(service: Service): Promise<void>;
    assignCallerUserRole(user: Principal, role: UserRole): Promise<void>;
    deleteContact(id: string): Promise<void>;
    deleteService(id: string): Promise<void>;
    getAllContacts(): Promise<Array<ContactSubmission>>;
    getAllServices(): Promise<Array<Service>>;
    getCallerUserProfile(): Promise<UserProfile | null>;
    getCallerUserRole(): Promise<UserRole>;
    getService(id: string): Promise<Service>;
    getUserProfile(user: Principal): Promise<UserProfile | null>;
    isCallerAdmin(): Promise<boolean>;
    reorderServices(orderedServiceIds: Array<string>): Promise<void>;
    saveCallerUserProfile(profile: UserProfile): Promise<void>;
    submitContact(contact: ContactSubmission): Promise<void>;
    updateService(service: Service): Promise<void>;
}